Give the stall monitor vocabulary types their own files
Aziz one-major-type policy: TerminalRenderDropGate,
TerminalStallRecoveryCause, TerminalStallEmission, and
TerminalRenderStallSnapshot are independent public vocabulary types and
now live in their own files; TerminalRenderStallMonitor.swift keeps
only the monitor and its private episode bookkeeping.

Part of #7573
