+++ Packages/Shared/CMUXMobileCore/Sources/CMUXMobileCore/TerminalRenderStallMonitor.swift
@@ -1,73 +1,5 @@
 import Foundation
 
-/// A render-path gate that can intentionally drop terminal frames.
-public enum TerminalRenderDropGate: UInt8, Sendable, Codable, CaseIterable {
-    /// Output is behind a pending terminal input acknowledgement sequence.
-    case pendingInputSeq = 1
-    /// A terminal replay barrier is active and live output is being held back.
-    case replayBarrier = 2
-    /// Render-grid output is waiting for a full baseline frame.
-    case baselineWait = 3
-    /// A viewport resize barrier is active while geometry is being reconciled.
-    case viewportBarrier = 4
-
-    /// The bit used in compact gate-active snapshots.
-    public var bit: Int {
-        1 << Int(rawValue - 1)
-    }
-}
-
-/// What resolved a previously detected terminal render stall.
-public enum TerminalStallRecoveryCause: UInt8, Sendable, Codable, CaseIterable {
-    /// A later frame caught the surface up and rendered normally.
-    case catchupFrame = 1
-    /// A replay acknowledgement cleared the barrier.
-    case replayAck = 2
-    /// A terminal output resync was triggered.
-    case resync = 3
-    /// The user explicitly requested a refresh or render reset.
-    case manualRefresh = 4
-    /// The user or network path forced a reconnect.
-    case reconnect = 5
-    /// The surface detached, so the episode is no longer relevant.
-    case surfaceDetached = 6
-    /// A barrier cleared for a non-ack reason.
-    case barrierCleared = 7
-}
-
-/// A stall monitor emission that callers can record or upload.
-public enum TerminalStallEmission: Sendable, Equatable {
-    /// A drop episode crossed the configured stall threshold.
-    case stallDetected(
-        surface: UInt32,
-        gate: TerminalRenderDropGate,
-        droppedFrames: Int,
-        stallDuration: TimeInterval
-    )
-    /// A previously detected stall episode resolved.
-    case stallRecovered(
-        surface: UInt32,
-        gate: TerminalRenderDropGate,
-        how: TerminalStallRecoveryCause,
-        duration: TimeInterval,
-        droppedFrames: Int
-    )
-}
-
-/// A point-in-time view of an open render drop episode.
-public struct TerminalRenderStallSnapshot: Sendable, Equatable {
-    /// The surface the episode belongs to.
-    public let surface: UInt32
-    /// The gate currently dropping frames.
-    public let gate: TerminalRenderDropGate
-    /// Number of dropped frames observed in the episode.
-    public let droppedFrames: Int
-    /// How long the episode has been open.
-    public let duration: TimeInterval
-    /// Whether ``TerminalStallEmission/stallDetected(surface:gate:droppedFrames:stallDuration:)`` already emitted.
-    public let stallDetected: Bool
-}
-
 /// Tracks terminal render drop episodes without timers or shell dependencies.
 ///
 /// Detection is drop-driven: a gate that keeps dropping frames crosses the
